clara-backend/requirements.txt
# Clara Multi-Agent Backend - Requirements
# Python 3.10+

# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# AI/LLM
openai>=1.3.0
anthropic>=0.7.0
groq>=0.4.0

# Database
supabase>=2.0.0

# Utilities
httpx>=0.25.0
python-multipart>=0.0.6

# Logging & Monitoring
loguru>=0.7.0
colorama>=0.4.6

# Validation
email-validator>=2.0.0
phonenumbers>=8.13.0

